executor: move tidb_reader code to its named files.  (#7065)
